+++ Runtime/AssemblyInfo.cs
@@ -1,2 +1,3 @@
 using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.2D.PixelPerfect.Tests")]
+[assembly: InternalsVisibleTo("Unity.2D.PixelPerfect.Editor")]